Resolve cyclical import and move sensorClasses to sensor
